Fix transaction reversal

+++ src/main/java/com/minibank/demo/user/ITransaction.java
@@ -1,13 +1,14 @@
 package com.minibank.demo.user;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public interface ITransaction {
     void process() throws Exception;
     void setBankAccount(BankAccount bankAccount);
-    void setAmount(Double amount) throws Exception;
+    void setAmount(BigDecimal amount) throws Exception;
     void setId(Long id);
     LocalDateTime getTimestamp();
     String getType();
-    Double getAmount();
+    BigDecimal getAmount();
 }
